requirements.txt
# Core Dependencies
pyspark==3.5.0
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
mlflow==2.9.2

# Data Validation & Quality
great-expectations==0.18.8

# Dashboard
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0

# Database & Storage
delta-spark==3.0.0
pyarrow==14.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0
jupyter==1.0.0

# Azure (if using Azure Databricks)
azure-storage-blob==12.19.0
azure-identity==1.15.0